Connect the roll result and button to the view model.
This gets us to a working app. It still has the files Intellij made and
things like the floating action button that we're not using.

The next steps are to create a settings fragment to control the number
of dice and the number of sides. Some tests would be good too.