Create an instrumented test.
This commit is primarily a demonstration of getting an instrumented test
working when a) you're using Hilt and b) you're testing a fragment, as
there seems to be a shortage of Java examples of both.

The HiltTestActivity gets added to the new debug source set so that it
isn't included in the "real" release build.